backend/requirements.txt
fastapi
pydantic
pydantic-settings
requests
apprise
python-dotenv
loguru
paho-mqtt
apscheduler
pytz
sqlalchemy
